Splits SplittableParDo into a core-construction part and a KWI-related part
